Move debugging facilities to D.HM.I.Debug
